+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Added tests for whitespace handling.
   - Added tests for entity decoding.
   - Added tests for node functions after calling remove().
+  - Added tests for `maxLen` in `file_get_html`.
   - Added test for bug [#172](https://sourceforge.net/p/simplehtmldom/bugs/172/)
 - Added optional argument `$trim = true` to `$node->text()`
 - Added attribute value normalization